requirements.txt
openai==1.17.1
ipykernel==6.29.3
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3
langchain==0.1.13